requirements.txt
streamlit
torch
torchvision
efficientnet-pytorch
pillow
matplotlib
numpy